Major Changes - Collaborative movie recommendation system is complete in Data_processing.py. 2 more changes below:
Moved the database initialisation into database.py so it can be accessed more easily in other files.
App.py has the function call for now. When the website is up and running this will change such that a new user's preferences can be input into the database and then recommendations can be provided