MAJOR CHANGES - Structural changes, web page is now becoming functional
Structure: the structure of the code for the project has change into a single python class. When the program is run, it runs any web pages from the classes methods and additional functions are also included in these methods.
This should improve the readability of the code.

Web page: the web page has been functional. Now the selection forms are being coded.