Script created for the home.html. @/static/script.js
Script: the script now generates a list of anime that the user can submit their preferences for. It dynamically adjusts between 5 and 10 depends on what the user prefers.

Next step, including the name data from flask into this process, such that each option has another drop down with anime names. The selections and their ratings will be passed back to the app where it will be processed